Moons rotation duration

+++ src/generation/generator.h
@@ -48,7 +48,7 @@ public:
 	Q_INVOKABLE int		moonCount(qreal planetSize);
 	Q_INVOKABLE qreal	moonSize(qreal planetSize);
 	Q_INVOKABLE qreal	moonOrbitalDistance(qreal planetSize, qreal moonSize);
-	Q_INVOKABLE	qreal	moonRotationDuration(qreal planetSize, qreal moonSize, qreal moonOrbitalDisance);
+	Q_INVOKABLE	int moonRotationDuration(qreal planetSize, qreal moonSize, qreal moonOrbitalDisance);
 
 protected:
 	std::mt19937 m_gen;
